+++ mxops/analyze/__init__.py
@@ -1,0 +1,5 @@
+"""
+author: Etienne Wallet
+
+This subpackage is used to record the parameters / attributes of the deployed contracts
+"""